to have less line changes
